feat: post list, add, edit and delete

+++ database/migrations/2025_02_17_122613_create_posts_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
                 ->cascadeOnDelete();
             $table->string('title');
             $table->string('description');
-            $table->decimal('price');
+            $table->unsignedBigInteger('price');
             $table->integer('quantity');
             $table->integer('quantity_left');
             $table->string('status');
